int to float

+++ core/src/com/gdxtest/EmptyBlock.kt
@@ -3,6 +3,6 @@ package com.gdxtest
 import com.badlogic.gdx.graphics.g2d.SpriteBatch
 
 class EmptyBlock : IBlock {
-  override fun show(batch: SpriteBatch, x: Int, y: Int) {
+  override fun show(batch: SpriteBatch, x: Float, y: Float) {
   }
 }
