+++ core/src/com/gdxtest/IBlock.kt
@@ -3,5 +3,5 @@ package com.gdxtest
 import com.badlogic.gdx.graphics.g2d.SpriteBatch
 
 interface IBlock {
-  fun show(batch: SpriteBatch, x: Int, y: Int)
+  fun show(batch: SpriteBatch, x: Float, y: Float)
 }
